Add support for user overriding NuGet.exe from package using NUGET_PATH system environment variable.
A version of NuGet.exe is downloaded as part of the pip install process but if that failed the code attempted to find NuGet.exe on the users path.  That logic never worked as the path was always invalid according to os.path.isfile.   Remove PATH based logic and instead allow the user to override the NuGet.exe version used by setting NUGET_PATH to the folder containing NuGet.exe.  If this environment variable is set it will always be used.  If NuGet.exe doesn't exist in the folder defined by NUGET_PATH then the operation will fail.   

Fixes #259

Co-authored-by: Brandon Huber <brandon.huber@hp.com>