Integration Tests - Update scripts to programmatically find default branch from remote. (#276)
By finding the default branch of a repo automatically this will avoid having to update these tests each time a remote repo changes their default branch.  This also ensures the branch being tested is the most current.

For Project Mu this happens a lot based on the release process.  
For edk2 usage this currently isn't a problem but this change should make it more resilient to future changes.