Bug: Check all directories in env directory traversal (#290)
* Bug: Check all directories in env directory traversal

161e4be added support to self_describing_environment._gather_env_files()
to skip directories specified by a platform. That commit reused existing
directory walk logic which modifies the dir list used for the walk based
on index.

This does not work when multiple directories might be removed
from a single root as when a directory is removed in-place during
enumeration, the index is incremented by one and the next directory is
removed. This effectively causes the new index to skip the actual next
item on the next iteration.

For example:
  dirs = ['.git', 'a', 'b', 'c', 'd']
  i = 1
  del dirs[i] removes dirs[1] => 'a'

  dirs = ['.git', 'b', 'c', 'd']
  i = 2
  dirs[i] now = 'c' not 'b'

This change modifies the in-place assignment of dirs to use slice
assignment which similarly prunes the dir list for subsequent file
enumeration without depending upon an index value.

This would not surface as an issue unless two consecutive directories
in the directory are list, in which case, the second directory is
skipped.

test_multiple_duplicate_ext_deps_skip_dir() added to test_edk2_update.py
demonstrates this. It fails with the previous logic as
six/seventh/eighth directly proceeds third/fourth/fifth causing it to
be skipped over. It is properly removed with the new logic.