Update usage of deprecated loading APIs
Fixes #337

Plugins are currently loaded in the `_load()` function in
plugin_manager.py by using the `imp` standard library:

https://docs.python.org/3/library/imp.html

The first sentence of the library documentation states the following:

```
Deprecated since version 3.4, will be removed in version 3.12: The
imp module is deprecated in favor of importlib.
```

This means, `imp` will be removed in the next Python release.

Therefore, this change updates the code to use `importlib` standard
library as recommended by the `imp` documentation.

The code follows the recipe recommended by the `importlib`
documentation to directly load a Python source file as is the case
when loading Python modules based on plugin descriptor information.

https://docs.python.org/3/library/importlib.html#importing-a-source-file-directly

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>