Add required python version for pypi
Adds a required minimum version of python during pypi release that is
mirrored on pypi.org so that it is obvious what python versions are
supported for edk2-pytool-extensions.

Additionally adds a new contributor doc file that specifies all files
that must be updated when a new minor version of python is released.