Respect git worktrees in the SDE
Closes #367

Git worktrees are a standard, well-defined git feature. See:
  https://git-scm.com/docs/git-worktree

The Self-Describing Environment (SDE) currently does not recognize
git worktrees which means it fails with multiple descriptors present
since the worktree is a copy/variation of the base repo.

The SDE should recognize git worktrees and not load descriptors from
the worktrees if they are not the active workspace.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>