Update edk2_git.py to use PyGit2
edk2_git.py originally relied on calling git commands directly with
RunCmd to get status information about a repo. This was slow and due
to the implementation, was static (would only be called once, not
updating status if the repo was updated).

This change updates edk2_git.py to use PyGit2, which parses the git
database to get status information about the repo. As a side effect,
this information is now dynamic (If the repo is updated, the status is
updated).

Signed-off-by: Joey Vagedes <joeyvagedes@microsoft.com>