Nuget Publishing deprecation: switch from LicenseUrl to License (#375)
Nuget license requirements have been updated such that when uploading
a package, a license xml element must be used that specifies that the 
license is a typical expression (i.e. BSD2, BSD3, etc), or a file uploaded
with the package. This commit does the following:

Updates nuget-publish (nuget_publishing.py) 'New' Command to support
the new Nuget license requirements when generating the config file.
The user must provide the '--LicenseIdentifier' argument and a common
license type or the '--CustomLicense' flag to specify a custom license
will be provided in during the 'Pack' Command.

Updates nuget-publish (nuget_publishing.py) 'Pack' Command to support
the new Nuget license requirements when reading the config file by
generating the nuspec file with the license xml element that specifies
that the license is an expression (typical license) or an uploaded file.
If the file is being uploaded, the absolute path must be provided via the
'--CustomLicensePath' argument.

Signed-off-by: Joey Vagedes <joeyvagedes@microsoft.com>