Add CLI env information to the argument parser --help
Adds a method to Edk2Invocable, AddParserEpilog(), that allows an
invocable to customize the epilog of the argument parser when --help is
used. In particular, Edk2PlatformBuild uses this method to append
platform specified critical build env variables so that developers
no longer need to dig through the code to see what build customizations
are available.

Signed-off-by: Joey Vagedes <joeyvagedes@microsoft.com>