Add ability to skip verification of external dependencies (#419)
Adds two separate ways to skip verification of external dependencies. This is to enable co-development of feature repos that are downloaded as external dependencies instead of submodules.

Adds a --noverify flag to stuart_build which will skip verification for all external dependencies for that single run
Adds a new state flag in the for an external dependency's state file (extdep_state.yaml), verify, that by default is true, but can be set to false to disable verification indefinitely for that external dependency.