Improve CI Testing performance
Add pytest-xdist to parallelize the tests across threads and CPUs if
applicable.

Set dependabot to only update once a week. This improves performance as
we cache pip dependencies to be used in future CI, however with
dependencies being every day, much of the usefulness of this feature is
being lost, and we are just storing an excess amount of dependency
caches.

update dependency name

Update

Update

Update

Update

Update

Update

Improve CI Testing performance

Add pytest-xdist to parallelize the tests across threads and CPUs if
applicable.

Set dependabot to only update once a week. This improves performance as
we cache pip dependencies to be used in future CI, however with
dependencies being every day, much of the usefulness of this feature is
being lost, and we are just storing an excess amount of dependency
caches.

update dependency name

Update

Update

Update

Update

Update

Revert simple changes

Update to include workspace root

Update

u