Update to project to use black for formatting
Updates the necessary files to use black as the default formatter for
the project. This includes updating vscode settings to recommend two
extensions and then enables autoformatting on save, using those
extensions.

Also updates the CI to run black on the project and fail if there are
any changes that need to be formatted. Additionally updates the
developing documentation to include instructions on how to use black.

Signed-off-by: Joey Vagedes <joeyvagedes@microsoft.com>