repo_resolver: add git version Check (#486)
Previously, the git version only mattered to edk2_setup, so the check was appropriate there. Now that the git version matters to much of the build system due to it's use of the -z flag when reading git worktrees, the check needed to be moved to a more appropriate place. This commit moves it from edk2_setup to repo_resolver.py:get_details()