Remove `GetNugetCmd()` from setup
With recent changes to pip, using a method from the package during the
setup of the package is no longer supported. Due to this, it was decided
to move when we download NuGet to when we first attempt to use it. This
has multiple benifits as we now no longer download NuGet, even if the
user is supplying their own version of NuGet, and our download NuGet
functionality remains inside the package and easily testable, rather
then an extra function inside of setup.py

Additionally we no longer use the deprecated module pkg_resources to
identify the binary folder inside the package, instead using
importlib's resources module to do this.