edk2_logging.py: increase specificity of NuGet API key (#513)
The False positive rate of API keys being filtered out was a detriment to developer productivity per #502. This commit reduces the API key to be only lowercase a-z as NuGet API keys generated from NuGet.org are generated as all lowercase. This should reduce the amount of false positives noted, specifically false positives due to the long folder names in the build output folder.