external_dependency: add sha256 verification
Previously, the default implementation only verified the version, and
did not verify files within the directory had not changed. This type of
functionality did exist for git dependencies and for web dependencies
as long as it was a single file. This commit adds sha256 verification
for all dependencies by default.

This is completed by calculating the sha256 hash for a dependency if it
is not provided in the extdep file. This calculation works by finding
all files in the content directory, sorting them, then hashing them
chunk by chunk. This hash is stored in the state file.

Then, whenever the SDE verifies the environment, the hash is calculated
again and compared.