repo_resolver: bugfix error printing
When resolving a repository dependency, if the URL of the existing
repository does not match the URL of the actual dependency, an error
is printed to the user stating there is a mismatch. A bug was found
in the code that prints this error. The attribute trying to be accessed
did not exist because there was an interface change in how that
information is accessed.

This bugfix updates the logic to access the information the correct way.