Update txt logger mode to 'a' (#584)
Python logging to a file with mode='w' can cause log loss when logging happens after the logger has already been cleaned up.
See issue: python/cpython#86544

The fix for Python was to have special cases for files opened in mode='w'. Currently we use mode='w+', so the checked in fix does not solve our problem.

I've implemented moving the file_logger to mode='a' and an explicit deletion of the logging file before opening the FileHandler.