nuget_dependency: authorization failure logging
When a nuget package fails to download due to an authorization error,
additional logging is provided to help the user understand the problem
and possible solutions. The message informs the user to either install
a credential provider (because the the package requires authorization)
or to run the command themselves to provide the credentials manually.