Deprecate multiple_workspace: MultipleWorkspace
MultipleWorkspace has many similarities to the Edk2Path object and only
the .join() function is used internal to the edk2toolext package.
Edk2Path's GetAbsolutePathFromEdk2RelativePath() function can be used as
a drop in replacement for the .join() function, resulting in an easy to
replace deprecation.

Looking at the usage of MultipleWorkspace by
external consumers also show that they are only using the .join()
function, however the other functions can also be easily replaced with
Edk2Path functions.