Simple heuristics to try and locate platform module
Currently, we only support not providing the platform module (-c) if
you are in the correct directory and it is a called PlatformBuild.py

This commit slightly expands the heuristic for locating a platform
module to also include CISettings.py and .pytool/CISettings.py AND it
will search parent directories for those three possible locations. If
one is found, it will be used. Otherwise it will print the same message
it already prints.