coverage_report: Allow package filtering for --by-platform
Originally, both a --by-package and --by-platform flag was created to
describe the types of coverage reports able to be generated, and allow
for grouping the command line arguments based on these two flags.

With the adding of the ability to filter on packages for the platform
report, the --by-package flag no longer has any special arguments and
can be deprecated. Due to this, this change also removes the
--by-package flag, making it the default behavior, which is overridden
by --py-platform.