edk2_invocable: Bugfix for bad argument handling
This change fixes a bug in the edk2_invocable.py script where an invalid
command line argument would not cause the script to exit with an error.

This occured due to a change that enabled non-valued environment
variables to be passed to the script. The result of this change was that
any invalid command line argument would be added as a non-valued
environment variable rather than triggering the error handling.

This change is such that any non-valued environment variable can not
start with a '-' character, as that is reserved for argparse arguments.