nuget_dependency: Allow path override (#871)
As described in [external_dependency.py#L38](https://github.com/tianocore/edk2-pytool-extensions/blob/1e4ddee16145c262b96c67bd442124563400d8d6/edk2toolext/environment/external_dependency.py#L38), the `name` configuration is the name of the ext_dep, used to name the folder the ext_dep will be unpacked into. When the nuget external dependency was added, `name` also became the name of the package you are downloading, which tightly couples the name of the package and the folder path. This means there is no way to name the folder something other than the package name, which can be necessary in some situations.

This change provides a backwards compatible way to de-couple the package name from the folder name by introducing a new, optional config for nuget dependencies, `package`. `package` will be used as the package name to download while `name` will continue to be the name of the ext dep and the folder it is downloaded to. For backwards compatability, if `package` is not used in the configuration file, `name` will be used instead.