Python support: drop 3.10, add 3.13 (#895)
## Breaking Change

This commit removes support for python 3.10. Attempting to install v0.28.0 of edk2-pytool-extensions using a python version less than 3.11 will result in an error from pip along the lines of `ERROR: Could not find a version that satisfies the requirement ...`

## Integration Instructions

If your platform cannot upgrade to 3.11 or above, continue using a version of edk2-python-extensions less than v0.28.0. Otherwise upgrade your version of platform to 3.11 or greater. If edk2-pytool-library is not automatically downloaded with edk2-pytool-extensions, then you also need to ensure that edk2-pytool-library is v0.22.0 or greater