improve verbosity flag override (#938)
Improves the verbosity command line argument for stuart commands in the following ways:

Moves when the flag is registered, so that the flag can be seen with the --help flag, even if a configuration file is not used.
Passes the verbosity flag to the edk2 build command.
Additionally, as a part of (2), platforms can now also control verbosity of EDKII Build in their platform configuration files via the EDK2_BUILD_VERBOSE env variable.