Improve configuration options for GetDependencies. (#966)
GetRequiredSubmodules() vs GetDependencies()

GetRequiredSubmodules() allows a recursive/omnicache configuration parameter to be passed. 
The configuration parameter determines if --recursive is passed when the submodule is cloned.

GetDependencies() does not have configuration for this type of scenario.
A Dependency from GetDependencies will always attempt to resolve **ALL** submodules that are contained in the repo.

This presents a problem when a submodule contains their own dependency as well. The example run into is:

SecurityPkg => libspdm
libspdm => openssl
openssl => gost-engine
gost-engine => libprov

```
{
	"Path": "Common/MU_TIANO",
	"Url": "https://github.com/microsoft/mu_tiano_plus.git",
	"Branch": "dev/202502",
},
```

```
{
	"Path": "Common/MU_TIANO",
	"Url": "https://github.com/microsoft/mu_tiano_plus.git",
	"Branch": "dev/202502",
	"Recurse": True
},
```

```
{
	"Path": "Common/MU_TIANO",
	"Url": "https://github.com/microsoft/mu_tiano_plus.git",
	"Branch": "dev/202502",
	"Recurse": False
},
```
```
{
	"Path": "Common/MU_TIANO",
	"Url": "https://github.com/microsoft/mu_tiano_plus.git",
	"Branch": "dev/202502",
	"Recurse": {"CIFile" : ".pytool/CISettings.py"}
},
```

Add "Recurse" support for GetDependencies function.
Possible values are True/False/Dictionary that Specifies a CIFile.
If a CI file is specified, the GetRequiredSubmodules of the CI file will be
used to determine which submodules should be cloned.