environment/repo_resolved: Support branches from gh pr checkout
When checking out a PR using the gh command, the tracking
branch may not be associated with any of the current remotes.
This causes a python exception determining the URL of the
remote associated with the PR branch when running stuart
commands.

However, the URL to the repo the PR was created from can be
determined by parsing the tracking_branch() and removing the
branch name from the end of the tracking_branch().

Signed-off-by: Michael D Kinney <michael.d.kinney@intel.com>