edk2_invocable: Register one version per pip package (#989)
This commit results in only the first version of a pip package found on the path being used as the version, even if multiple versions exist in the distribution list because it exists in multiple path locations. If using a venv environment, the path to those pip modules is prepended, so it should theoretically always be correct.