invocable: Do not print help message on error. (#1002)
The help message is extremely large for our invocables, and tends to
cause the terminal to scroll up, which ultimately causes the actual error
message to be lost in the noise.

Instead, lets continue to print the error message, but only add a note
that the user can use `--help` to print the help message if they wish.