Format output xml and align GUID format for fpdt_parser (#1006)
Format output xml and align GUID format for fpdt_parser.
Apply `minidom` to format the output xml of the `fpdt_parser.py` to improve readability.
The script was not adding entries to FBDT table so the output is not a proper tree structure.
Before it was:
```
|- fbpt_header
|- fbpt_record
|- fbpt_record
```
Now it is:
```
|- fbpt_header
|   |- fbpt_record
|   |- fbpt_record
```
it can be properly consumed by `perf_report_generator.py` to generate perf report. Before this change there will be 0 entries in the report because the input structure doesn't match what the script needs.

Update the GUID format for different entries to make sure the output GUID format stay the same and can be recognized by the perf_report_generator.py.
Add xml formatter to make sure the output xml more readable.

Examples for GUID format change:
```
Before this fix:
For DynamicStringEvent: 
            `<GUID Value="23C9322F-2AF2-476A-BC4C-26BC-8826-6C71"/>`
For DualGuidStringEvent: 
            `<GUID1 Value="6D33944A-EC75-4855-A54D-809C75241F6C"/>`
            `<GUID2 Value="7B94C75C-36A4-4AA4-A1DF-14BC9A049AE4"/>`

After this fix:
All event guids will be in the same format of 
            `<GUID1 Value="6D33944A-EC75-4855-A54D-809C75241F6C"/>`
```
With this change the `perf_report_generator.py` can successfully parse these GUIDs and map to the right module name.

- [x] Breaking change?
Detail of the changes can be seen above. The structure and content of the output xml for `fpdt_parser.py` changed.

- [x] Integration instruction:
Use `fpdt_parser.py` first and use the output xml file as the input for the `perf_report_generator.py` to get perf report.