web_dependency: Do not write zerod file permission
When decompressing a zip file on a linux system, we attempt to restore
the previous file permission so that executables stay executable. A bug
was found such that if the underlying files were created on a Windows
system and then zipped, they do not contain file permissions, so when
restoring the previous file permissions, we were "chmod"ing "0".

Different Linux distributions handle that differently. One was found to
completely wipe all permissions, which prevents any user from moving the
file, which is what the next step in the web_dependency process is.

The current solution, in this commit, is to only apply the permissions
if they exist (e.g. they are not zero). If they do not exist, then we
will just keep the OS provided permissions that come with a newly
created file.