web_dependency: Temporarily disable test
The test that downloads a 4.7MB file from GNU FTP is pretty reliably
dropping the connection 3/4 of the way through the process of
downloading and writing the contents to a local temporary file. This
does not appear to be an issue with our code, as I have tested with
multiple web request libraries (urllib, urllib3, requests) and received
the same underlying error every time. I've also trying reading by chunks
instead of the entire file at once.

At this point, I believe the issue is stemming from a configuration on
the server side, rather than in our code. An issue will be created for
this test and it will be investigated further, but it is currently
blocking an important PR from merging. This will be reverted when the
issue is resolved.