edk2_parse: bugfix unknown arguments for platform
For edk2_parse to accurately parse the workspace, it must set up it's
environment (the environment variables - e.g. the variables that can be
accessed via `self.env.GetValue(...)`). One way it supports doing this
is associating itself with a platform build variable and calling it's
pre-build setup.

This includes calling the platform build's `RetrieveCommandLineOptions`
function. However there was a bug in that logic in which we do not also
call the platform's `AddCommandLineOptions` function. Due to this, if
the platform has a custom build argument, it won't be available when
`RetrieveCommandLineOptions` is called, ending in an error `'Namespace'
object has no attribute '<arg_name>'`.

This commit fixes this bug by first calling `AddCommandLineOptions` on a
new parser, re-parsing the environment, and passing all known and parsed
args to `RetrieveCommandLineOptions`.